Split "create_seed" into "create" and "lint" commands.
