+++ src/scripts/lint.ts
@@ -53,8 +53,7 @@ function getLintDiffCommands(options: Options): Record<string, any> {
     '*.{ts,js,tsx,jsx}':
       'eslint --config src/.eslintrc.js' + (options.fix ? ' --fix' : ''),
     'studies/*': () =>
-      'npm run seed_tools -- create_seed studies --validate_only' +
-      (options.fix ? ' --fix' : ''),
+      'npm run seed_tools -- lint studies' + (options.fix ? ' --fix' : ''),
   };
 }
