+++ src/seed_tools/commands/create.ts
@@ -1,0 +1,66 @@
+// Copyright (c) 2024 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import { Command } from '@commander-js/extra-typings';
+import { createHash } from 'crypto';
+import { promises as fs } from 'fs';
+import { VariationsSeed } from '../../proto/generated/variations_seed';
+import { readStudiesToSeed } from '../utils/studies_to_seed';
+
+export default function createCommand() {
+  return new Command('create')
+    .description('Create seed.bin from study files')
+    .argument('<studies_dir>', 'path to the directory containing study files')
+    .argument('<output_seed_file>', 'output seed file')
+    .option('--mock_serial_number <value>', 'mock serial number')
+    .option(
+      '--output_serial_number_file <path>',
+      'file path to write the seed serial number',
+      './serialnumber',
+    )
+    .option('--version <value>', 'seed version to set')
+    .action(createSeed);
+}
+
+interface Options {
+  mock_serial_number?: string;
+  output_serial_number_file: string;
+  version?: string;
+}
+
+async function createSeed(
+  studiesDir: string,
+  outputSeedFile: string,
+  options: Options,
+) {
+  const { variationsSeed, errors } = await readStudiesToSeed(studiesDir, false);
+
+  if (errors.length > 0) {
+    console.error(`Seed validation errors:\n${errors.join('\n---\n')}`);
+    process.exit(1);
+  }
+
+  const serialNumber =
+    options.mock_serial_number ?? generateSerialNumber(variationsSeed);
+  variationsSeed.serial_number = serialNumber;
+
+  variationsSeed.version = options.version ?? '1';
+
+  console.log('Seed study count:', variationsSeed.study.length);
+  const seedBinary = VariationsSeed.toBinary(variationsSeed);
+  await fs.writeFile(outputSeedFile, seedBinary);
+  await fs.writeFile(options.output_serial_number_file, serialNumber);
+  console.log(outputSeedFile, 'created with serial number', serialNumber);
+}
+
+function generateSerialNumber(variationsSeed: VariationsSeed): string {
+  const seedBinary = VariationsSeed.toBinary(variationsSeed);
+  const timestamp = String(Date.now());
+  const hash = createHash('md5')
+    .update(seedBinary)
+    .update(timestamp)
+    .digest('hex');
+  return hash;
+}
