+++ src/seed_tools/seed_tools.ts
@@ -6,13 +6,15 @@
 import { program } from '@commander-js/extra-typings';
 
 import compare_seeds from './commands/compare_seeds';
-import create_seed from './commands/create_seed';
+import create from './commands/create';
+import lint from './commands/lint';
 import split_seed_json from './commands/split_seed_json';
 
 program
   .name('seed_tools')
   .description('Seed tools for manipulating study files.')
   .addCommand(compare_seeds())
-  .addCommand(create_seed())
+  .addCommand(create())
+  .addCommand(lint())
   .addCommand(split_seed_json())
   .parse();
